Fix wonky Darwin linker test in configure.ac
On Ubuntu 12.04, this check for the Darwin linker was passing, when it should not have. This resulted in a link failure when the build system tried to pass the Darwin-specific (?) `-compatibility_version` flag to the linker.

My autotools knowledge is practically non-existent, but this patch seems to make the test pass on OS X and fail on Linux, which I'm assuming is what's supposed to happen.