Work around issue when calling dispatch_io_read() with `size` of 0
In some cases, if you pass dispatch_io_read() a size of zero, the I/O handler and the channel's cleanup handler execute concurrently.
This test requires the cleanup handler to execute after the I/O handler has returned for the last time (it sets `is_done` and `d`) i
so we really don't want this.